Added DATA folder but still haven't moved files
