+++ SCRIPTS/Thesis-plots.py
@@ -2,7 +2,7 @@
 import time as time
 import numpy as np
 import UTILS.material_structure as ms
-import material_model as mm
+import UTILS.material_model as mm
 import scipy.optimize as optimize
 import matplotlib.pyplot as plt
 import SCRIPTS.data_structure as ds
